+++ src/runtimes/paseo.rs
@@ -105,7 +105,7 @@ pub async fn try_crunch(crunch: &Crunch) -> Result<(), CrunchError> {
     let seed_account_id: AccountId32 = signer_keypair.public_key().into();
 
     // Get signer account identity
-    let signer_name = get_display_name(&crunch, &seed_account_id, None).await?;
+    let (signer_name, _) = get_display_name(&crunch, &seed_account_id, None).await?;
     let mut signer_details = SignerDetails {
         account: seed_account_id.clone(),
         name: signer_name,
@@ -675,7 +675,7 @@ async fn collect_validators_data(
             Some(controller) => controller,
             None => {
                 let mut v = Validator::new(stash.clone());
-                v.name = get_display_name(&crunch, &stash, None).await?;
+                (v.name, v.has_identity) = get_display_name(&crunch, &stash, None).await?;
                 v.warnings = vec![format!("No controller bonded!")];
                 validators.push(v);
                 continue;
@@ -689,7 +689,7 @@ async fn collect_validators_data(
         v.controller = Some(controller.clone());
 
         // Get validator name
-        v.name = get_display_name(&crunch, &stash, None).await?;
+        (v.name, v.has_identity) = get_display_name(&crunch, &stash, None).await?;
 
         // Check if validator is in active set
         v.is_active = if let Some(ref av) = active_validators {
@@ -822,7 +822,7 @@ async fn get_display_name(
     crunch: &Crunch,
     stash: &AccountId32,
     sub_account_name: Option<String>,
-) -> Result<String, CrunchError> {
+) -> Result<(String, bool), CrunchError> {
     let api = crunch.client().clone();
 
     let identity_of_addr = node_runtime::storage().identity().identity_of(stash);
@@ -840,7 +840,7 @@ async fn get_display_name(
                 Some(child) => format!("{}/{}", parent, child),
                 None => parent,
             };
-            Ok(name)
+            Ok((name, true))
         }
         None => {
             let super_of_addr = node_runtime::storage().identity().super_of(stash);
@@ -860,7 +860,7 @@ async fn get_display_name(
                 .await;
             } else {
                 let s = &stash.to_string();
-                Ok(format!("{}...{}", &s[..6], &s[s.len() - 6..]))
+                Ok((format!("{}...{}", &s[..6], &s[s.len() - 6..]), false))
             }
         }
     }